requirements.txt
github-copilot-sdk>=0.1.0
geopy>=2.4.0
pydantic>=2.0.0
